@@ -4,7 +4,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this file,
 // You can obtain one at https://mozilla.org/MPL/2.0/.
 
-package util
+package structutil
 
 import (
 	"testing"
@@ -52,46 +52,6 @@ func TestSetDefaults(t *testing.T) {
 		t.Errorf("bool failed")
 	} else if x.E.Value != "defaulter" {
 		t.Errorf("defaulter failed")
-	}
-}
-
-func TestUniqueStrings(t *testing.T) {
-	tests := []struct {
-		input    []string
-		expected []string
-	}{
-		{
-			[]string{"a", "b"},
-			[]string{"a", "b"},
-		},
-		{
-			[]string{"a", "a"},
-			[]string{"a"},
-		},
-		{
-			[]string{"a", "a", "a", "a"},
-			[]string{"a"},
-		},
-		{
-			nil,
-			nil,
-		},
-		{
-			[]string{"       a     ", "     a  ", "b        ", "    b"},
-			[]string{"a", "b"},
-		},
-	}
-
-	for _, test := range tests {
-		result := UniqueTrimmedStrings(test.input)
-		if len(result) != len(test.expected) {
-			t.Errorf("%s != %s", result, test.expected)
-		}
-		for i := range result {
-			if test.expected[i] != result[i] {
-				t.Errorf("%s != %s", result, test.expected)
-			}
-		}
 	}
 }
 
@@ -145,83 +105,6 @@ func TestFillNillSlices(t *testing.T) {
 
 	if len(z.A) != 0 {
 		t.Error("length")
-	}
-}
-
-func TestAddress(t *testing.T) {
-	tests := []struct {
-		network string
-		host    string
-		result  string
-	}{
-		{"tcp", "google.com", "tcp://google.com"},
-		{"foo", "google", "foo://google"},
-		{"123", "456", "123://456"},
-	}
-
-	for _, test := range tests {
-		result := Address(test.network, test.host)
-		if result != test.result {
-			t.Errorf("%s != %s", result, test.result)
-		}
-	}
-}
-
-func TestCopyMatching(t *testing.T) {
-	type Nested struct {
-		A int
-	}
-	type Test struct {
-		CopyA  int
-		CopyB  []string
-		CopyC  Nested
-		CopyD  *Nested
-		NoCopy int `restart:"true"`
-	}
-
-	from := Test{
-		CopyA: 1,
-		CopyB: []string{"friend", "foe"},
-		CopyC: Nested{
-			A: 2,
-		},
-		CopyD: &Nested{
-			A: 3,
-		},
-		NoCopy: 4,
-	}
-
-	to := Test{
-		CopyA: 11,
-		CopyB: []string{"foot", "toe"},
-		CopyC: Nested{
-			A: 22,
-		},
-		CopyD: &Nested{
-			A: 33,
-		},
-		NoCopy: 44,
-	}
-
-	// Copy empty fields
-	CopyMatchingTag(&from, &to, "restart", func(v string) bool {
-		return v != "true"
-	})
-
-	if to.CopyA != 1 {
-		t.Error("CopyA")
-	}
-	if len(to.CopyB) != 2 || to.CopyB[0] != "friend" || to.CopyB[1] != "foe" {
-		t.Error("CopyB")
-	}
-	if to.CopyC.A != 2 {
-		t.Error("CopyC")
-	}
-	if to.CopyD.A != 3 {
-		t.Error("CopyC")
-	}
-	if to.NoCopy != 44 {
-		t.Error("NoCopy")
 	}
 }
 
