@@ -14,7 +14,7 @@ import (
 	"strconv"
 	"time"
 
-	"github.com/syncthing/syncthing/lib/util"
+	"github.com/syncthing/syncthing/lib/structutil"
 )
 
 type Report struct {
@@ -179,7 +179,7 @@ type Report struct {
 
 func New() *Report {
 	r := &Report{}
-	util.FillNil(r)
+	structutil.FillNil(r)
 	return r
 }
 
